Have implemented NLog to the project

+++ OnionArchitecrureProject/Controllers/BookController.cs
@@ -2,6 +2,8 @@
 using DomainLayer.DTO.BookDtos;
 using DomainLayer.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Hosting;
+using RepositoryLayer.Migrations;
 using ServiceLayer.Service.Contract;
 using ServiceLayer.Service.Implementation;
 using ServiceLayer.Service.UoW;
@@ -14,16 +16,18 @@ namespace WebAPI_Layer.Controllers
     {
         private IMapper _mapper;
         private IUnitOfWork _unitOfWork;
+        private readonly ILogger<BookController> _logger;
 
-        public BookController(IMapper mapper, IUnitOfWork unitOfWork)
+        public BookController(IMapper mapper, IUnitOfWork unitOfWork, ILogger<BookController> logger)
         {
             _mapper = mapper;
             _unitOfWork = unitOfWork;
+            _logger = logger;
         }
 
         [HttpGet]
         [Route("getall")]
-        public async Task<List<AllBooksDto>> GetAllBooks()
+        public async Task<IActionResult> GetAllBooks()
         {
             var repository = _unitOfWork.GetRepository<Book>() as BookService;
 
@@ -31,24 +35,34 @@ namespace WebAPI_Layer.Controllers
 
             var request = _mapper.Map<List<Book>, List<AllBooksDto>>(books.ToList());
 
-            return request;
+            _logger.LogDebug("Произведена выборка всех книг");
+
+            return StatusCode(200, request);
         }
 
         [HttpGet]
         [Route("get")]
-        public async Task<BookDto> GetBook(int id)
+        public async Task<IActionResult> GetBook(int id)
         {
             var repository = _unitOfWork.GetRepository<Book>() as BookService;
 
             var book = await Task.Run(() => repository.GetBookById(id));
 
+            if (book == null)
+            {
+                _logger.LogError($"Ошибка: Книга с id: {id} не найдена. Проверьте корректность ввода!");
+                return StatusCode(400, $"Ошибка: Книга с id: {id} не найдена. Проверьте корректность ввода!");
+            }
+
             var request = _mapper.Map<Book, BookDto>(book);
 
-            return request;
+            _logger.LogDebug("Выборка прошла успешно. Выбрана книга: " + book.Title);
+
+            return StatusCode(200, request);
         }
 
         [HttpPost("add")]
-        public async Task AddBook(AddBookDto book)
+        public async Task<IActionResult> AddBook(AddBookDto book)
         {
             var repository = _unitOfWork.GetRepository<Book>() as BookService;
             var newBook = _mapper.Map<AddBookDto, Book>(book);
@@ -56,10 +70,14 @@ namespace WebAPI_Layer.Controllers
             await Task.Run(() => repository.AddBook(newBook));
 
             _unitOfWork.SaveChanges();
+
+            _logger.LogDebug($"Книга: {book.Title} добавлена успешно.");
+
+            return StatusCode(200, $"Книга: {book.Title} добавлена успешно.");
         }
 
         [HttpPut("edit")]
-        public async Task UpdateBook(EditBookDto book)
+        public async Task<IActionResult> UpdateBook(EditBookDto book)
         {
             var repository = _unitOfWork.GetRepository<Book>() as BookService;
             var editBook = _mapper.Map<EditBookDto, Book>(book);
@@ -67,15 +85,32 @@ namespace WebAPI_Layer.Controllers
             await Task.Run(() => repository.UpdateBook(editBook));
 
             _unitOfWork.SaveChanges();
+
+            _logger.LogDebug($"Информация о книге: {book.Title} изменена успешно.");
+
+            return StatusCode(200, $"Информация о книге: {book.Title} изменена успешно.");
         }
 
         [HttpDelete]
-        public async Task DeleteBook(long id)
+        public async Task<IActionResult> DeleteBook(long id)
         {
             var repository = _unitOfWork.GetRepository<Book>() as BookService;
 
+            var book = await Task.Run(() => repository.GetBookById(id));
+
+            if (book == null)
+            {
+                _logger.LogError($"Ошибка: Книга с id: {id} не найдена. Проверьте корректность ввода!");
+                return StatusCode(400, $"Ошибка: Книга с id: {id} не найдена. Проверьте корректность ввода!");
+            }
+
             await Task.Run(() => repository.RemoveBook(id));
+
             _unitOfWork.SaveChanges();
+
+            _logger.LogDebug($"Книга с id: {id} успешно удалена из библиотеки.");
+
+            return StatusCode(200, $"Книга с id: {id} успешно удалена из библиотеки.");
         }
     }
 }
